Move document data for testing to a separate file
